Lessons/Lesson-02/pyproject.toml
[project]
name = "lesson-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
]
